+++ pkg/security/secl/compiler/eval/strings_test.go
@@ -9,8 +9,6 @@ package eval
 import (
 	"slices"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
 
 func TestStringValues(t *testing.T) {
@@ -292,19 +290,5 @@ func BenchmarkRegexpEvaluator(b *testing.B) {
 				b.Fatal("unexpected result")
 			}
 		}
-	})
-}
-
-func TestGetPublicTLD(t *testing.T) {
-	t.Run("valid-fqdn", func(t *testing.T) {
-		etldPlusOne := GetPublicTLD("www.yahoo.com")
-		assert.Equal(t, "yahoo.com", etldPlusOne)
-	})
-}
-
-func TestGetPublicTLDs(t *testing.T) {
-	t.Run("valid-fqdns", func(t *testing.T) {
-		etldPlusOnes := GetPublicTLDs([]string{"www.yahoo.com", "www.google.com", "ftp.yahoo.com"})
-		assert.Equal(t, []string{"yahoo.com", "google.com"}, etldPlusOnes)
 	})
 }
